+++ History.md
@@ -1,4 +1,8 @@
-0.0.5 / 2013-05-20
+0.0.7 / 2013-05-21
+==================
+  * Fixed relative folder bug
+  
+  0.0.5 / 2013-05-20
 ==================
   * Added config file, command line support, more